requirement.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
distlib==0.3.4
Django==3.2.10
django-rest-framework==0.1.0
djangorestframework==3.13.1
filelock==3.6.0
idna==3.3
Pillow==9.1.0
platformdirs==2.5.1
psycopg2-binary==2.9.3
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.8
virtualenv==20.14.1
